Fix requests
